hoisted states for shown cards & more button
